feat: migrate dataUseTranslation.spec
